Track cascade Css surface refinements
Adopt the renamed Cascade.Cursor / Cascade.Reader / Cascade.Error
modules, the record return shape of Css.of_string, and the renamed
rules_of_statements / custom_props_of_rules accessors. Bracket var
fallbacks now build their Syntax_fallback value from a Cursor.remaining
slice instead of the previous Css.syntax_fallback shortcut, and
custom_value_ident / custom_value_var_empty_fallback move under
Css.Variables. Tailwind tokens that lower to free-form CSS values
(animation, transform, transform-origin, mask-image) use
Css.Declaration.of_string rather than the dropped read_* helpers.

Utility-level effects of the new Css types: clip-path polygons carry
fill-rule and spacing flags, text-transform/text-indent gain a Case
and Indent wrapper, prose pseudo-elements emit single-colon form, the
typography clamp() simplifier is dropped (the value layer handles it),
and Var.string_of_kind_value bridges Css.kind to the new
Css.Properties.kind family.

+++ lib/prose.ml
@@ -147,10 +147,10 @@ let video = Css.Selector.element "video"
 
 (* Pseudo-element helpers that apply pseudo AFTER the where clause *)
 let with_before selector =
-  Css.Selector.compound [ selector; Css.Selector.Before Css.Selector.Double ]
+  Css.Selector.compound [ selector; Css.Selector.Before Css.Selector.Single ]
 
 let with_after selector =
-  Css.Selector.compound [ selector; Css.Selector.After Css.Selector.Double ]
+  Css.Selector.compound [ selector; Css.Selector.After Css.Selector.Single ]
 
 let radius len = Css.Radius { horizontal = [ Css.Length len ]; vertical = None }
